requirements/base.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-allauth==65.3.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-storages==1.14.4
django-anymail==12.0
Pillow==11.1.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
TexSoup==0.3.1
requests==2.32.3
python-magic==0.4.27
weasyprint==62.3
gunicorn==23.0.0
whitenoise==6.8.2
django-htmx==1.21.0
boto3==1.35.86
